Re-added CustomerRepository.cs
