+++ Amarath/DAL/Data/AmarathContext.cs
@@ -35,7 +35,7 @@ namespace Amarath.DAL.Data
         }
 
         //public DbSet<User> Users { get; set; }
-        //public DbSet<Customer> Customers { get; set; }
+        public DbSet<Customer> Customers { get; set; }
         public DbSet<User> Users { get; set; }
     }
 }
