+++ Amarath/Startup.cs
@@ -53,11 +53,11 @@ namespace Amarath
             });
 
             //TODO: Remove this SQL Example at the end
-           /* var optionsBuilder = new DbContextOptionsBuilder<AmarathContext>();
+            var optionsBuilder = new DbContextOptionsBuilder<AmarathContext>();
             var dbContext = new AmarathContext(optionsBuilder.Options);
             System.Diagnostics.Debug.WriteLine("\n\n\n****");
             var query = dbContext.Customers.First(x => x.LastName == "Liu");
-            System.Diagnostics.Debug.WriteLine("\n\n*** " + query.FirstName);*/
+            System.Diagnostics.Debug.WriteLine("\n\n*** " + query.FirstName);
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
